Made indexes for Decks, Comments and User" (#33)
Made Index for Comments when it searches for DeckId. Made Index for User when it searches for Username.
Made index for Deck when it searches for either is_public or user_id.
Also made some changes to a User database fields. So they are not vchar(max) but reduced to vchar(255) or similar.